Use a short domain
In f7774c2f9f15aaf8c25a47ce4d1fcd0f53ac7e72 the domain was changed to
be unique for every run, but this breaks somewhere. beaker-docker can't
log in to the machine. This changes the default to be example.com (which
it used to be), but keeps it as a parameter so it can be changed if
needed.

Fixes: f7774c2f9f15aaf8c25a47ce4d1fcd0f53ac7e72