Fix attaching modules form non-puppet namespace
VoxPupuli use the `puppet` namespace and its modules artifacts are named
puppet-*.tar.gz, but other organization have a different namespace and
using the task fail because the filename does not match the pattern.

The modulesync_config configuration ensure that the `pkg` directory is
ignored, so we can consider it to not be part of the module repository,
and at the time this command is run to only contain the latest build of
the module that we want to attach.